requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
PyYAML==6.0